backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
supabase-py>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
urllib3>=2.0,<3.0
pycryptodome>=3.20,<4.0

